requirements.txt
colorama==0.4.3
crc8==0.1.0
crcmod==1.7
HTMLTestRunner-Python3==0.8.0
PyYAML==5.3.1
random-dog==1.0.2
requests==2.24.0
selenium==3.141.0
sha1==1.2
splinter==0.13.0
urllib3==1.25.9
